# controllers/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Users and authentication
  description: A service that manages users and authentication to various services
  version: ""
host: localhost:8080
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /auth/clean/sessions:
    get:
      tags:
      - session
      summary: clean-sessions session
      description: Deletes all the sessions that have expired
      operationId: session#clean-sessions
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /auth/clean/token/login:
    get:
      tags:
      - session
      summary: clean-login-token session
      description: Cleans old login tokens from the database
      operationId: session#clean-login-token
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /auth/clean/token/merge:
    get:
      tags:
      - session
      summary: clean-merge-token session
      description: Cleans old account merge tokens from the database
      operationId: session#clean-merge-token
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
...
  /auth/logout/:session-id:
    post:
      tags:
      - session
      summary: logout-specific session
      description: Logout of a specific session
      operationId: session#logout-specific
      parameters:
      - name: session-id
        in: query
        required: true
        type: string
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /auth/logout/all:
    post:
      tags:
      - session
      summary: logout-other session
      description: Logout all sessions for the current user except their current session
      operationId: session#logout-other
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /auth/session:
    post:
      tags:
      - session
      summary: refresh session
      description: Take a user's session token and refresh it, also returns a new
        authentication token
      operationId: session#refresh
      parameters:
      - name: X-Session
        in: header
        required: true
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          headers:
            Authorization:
              type: string
            X-Session:
              type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
  /auth/sessions:
    get:
      tags:
      - session
      summary: get-sessions session
      description: Gets all of the sessions that are associated with the currently
        logged in user
      operationId: session#get-sessions
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/SessionGetSessionsResponseBody'
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /auth/token:
    post:
      tags:
      - session
      summary: redeemToken session
      description: Redeems a login token for credentials
      operationId: session#redeemToken
      parameters:
      - name: API-Key
        in: header
        required: false
        type: string
      - name: RedeemTokenRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/SessionRedeemTokenRequestBody'
          required:
          - token
      responses:
        "201":
          description: Created response.
          headers:
            Authorization:
              type: string
            X-Session:
              type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
...
definitions:
...
  SessionGetSessionsResponseBody:
    title: 'Mediatype identifier: all-sessions; view=default'
    type: object
    properties:
      currentSession:
        $ref: '#/definitions/SessionResponseBody'
      otherSessions:
        $ref: '#/definitions/SessionResponseBodyCollection'
    description: Get-SessionsResponseBody result type (default view)
    example:
      currentSession:
        browser: Et sunt.
        id: Ut cupiditate non.
        ip: Animi ullam.
        isMobile: false
        lastUsed: "2004-07-15T12:58:26Z"
        latitude: Rerum ad.
        location: Atque quod numquam doloribus nihil ut in.
        longitude: Blanditiis perferendis sint quo dolorem provident.
        mapUrl: Dolor omnis.
        os: Eum voluptatem recusandae aut quia non ex.
        userId: Et harum ut harum qui.
      otherSessions:
      - browser: Et sunt.
        id: Ut cupiditate non.
        ip: Animi ullam.
        isMobile: false
        lastUsed: "2004-07-15T12:58:26Z"
        latitude: Rerum ad.
        location: Atque quod numquam doloribus nihil ut in.
        longitude: Blanditiis perferendis sint quo dolorem provident.
        mapUrl: Dolor omnis.
        os: Eum voluptatem recusandae aut quia non ex.
        userId: Et harum ut harum qui.
      - browser: Et sunt.
        id: Ut cupiditate non.
        ip: Animi ullam.
        isMobile: false
        lastUsed: "2004-07-15T12:58:26Z"
        latitude: Rerum ad.
        location: Atque quod numquam doloribus nihil ut in.
        longitude: Blanditiis perferendis sint quo dolorem provident.
        mapUrl: Dolor omnis.
        os: Eum voluptatem recusandae aut quia non ex.
        userId: Et harum ut harum qui.
      - browser: Et sunt.
        id: Ut cupiditate non.
        ip: Animi ullam.
        isMobile: false
        lastUsed: "2004-07-15T12:58:26Z"
        latitude: Rerum ad.
        location: Atque quod numquam doloribus nihil ut in.
        longitude: Blanditiis perferendis sint quo dolorem provident.
        mapUrl: Dolor omnis.
        os: Eum voluptatem recusandae aut quia non ex.
        userId: Et harum ut harum qui.
  SessionRedeemTokenRequestBody:
    title: SessionRedeemTokenRequestBody
    type: object
    properties:
      User-Agent:
        type: string
        example: Nihil aut ut doloremque distinctio.
      token:
        type: string
        description: A merge token for merging into an account
        example: 38F93803-BFA0-82D1-0649-B0769F6F7519
        format: uuid
    example:
      User-Agent: Eius repellendus qui.
      token: AF3C2FF8-1A90-BD97-39C4-B4426C5778CC
    required:
    - token
  SessionResponseBody:
    title: 'Mediatype identifier: session; view=default'
    type: object
    properties:
      browser:
        type: string
        description: The browser and browser version connected with this session
        example: Dolorem et voluptate.
      id:
        type: string
        description: Unique unchanging session ID
        example: Rerum maxime in delectus quo ut eos.
      ip:
        type: string
        description: The last IP address where this session was used
        example: In assumenda dolor quam magni.
      isMobile:
        type: boolean
        description: Whether the session was from a mobile device
        example: false
      lastUsed:
        type: string
        description: Time that this session was last used
        example: "1979-08-24T10:02:47Z"
        format: date-time
      latitude:
        type: string
        description: The latitude of the last known location of the session
        example: Est accusantium sit.
      location:
        type: string
        description: A humanReadable string describing the last known location of
          the session
        example: Molestiae nihil adipisci facere ullam iure asperiores.
      longitude:
        type: string
        description: The longitude of the last known location of the session
        example: Facilis perferendis ipsam pariatur placeat expedita.
      mapUrl:
        type: string
        description: The URL of the Google map to show the location, suitable for
          using in an img tag
        example: Ut qui id cumque velit.
      os:
        type: string
        description: The OS of the system where this session was used
        example: Ratione dolorem nostrum quisquam voluptas.
      userId:
        type: string
        description: ID of the user this session is for
        example: Nemo aut et voluptatem autem earum.
    description: A session for a user, associated with a specific browser
    example:
      browser: Quisquam error recusandae totam.
      id: Itaque soluta exercitationem quod nihil repudiandae.
      ip: Odit unde.
      isMobile: true
      lastUsed: "2007-06-28T01:54:00Z"
      latitude: Sunt quia harum ut dolor repellat.
      location: Similique mollitia iure est non.
      longitude: Illum quo.
      mapUrl: Repellendus aliquam sit earum voluptatem natus.
      os: Accusantium quod.
      userId: Officia rerum.
    required:
    - id
    - userId
    - lastUsed
    - browser
    - os
    - ip
    - location
    - latitude
    - longitude
    - isMobile
    - mapUrl
  SessionResponseBodyCollection:
    title: 'Mediatype identifier: session; type=collection; view=default'
    type: array
    items:
      $ref: '#/definitions/SessionResponseBody'
    description: SessionCollectionResponseBody is the result type for an array of
      SessionResponseBody (default view)
    example:
    - browser: Et sunt.
      id: Ut cupiditate non.
      ip: Animi ullam.
      isMobile: false
      lastUsed: "2004-07-15T12:58:26Z"
      latitude: Rerum ad.
      location: Atque quod numquam doloribus nihil ut in.
      longitude: Blanditiis perferendis sint quo dolorem provident.
      mapUrl: Dolor omnis.
      os: Eum voluptatem recusandae aut quia non ex.
      userId: Et harum ut harum qui.
    - browser: Et sunt.
      id: Ut cupiditate non.
      ip: Animi ullam.
      isMobile: false
      lastUsed: "2004-07-15T12:58:26Z"
      latitude: Rerum ad.
      location: Atque quod numquam doloribus nihil ut in.
      longitude: Blanditiis perferendis sint quo dolorem provident.
      mapUrl: Dolor omnis.
      os: Eum voluptatem recusandae aut quia non ex.
      userId: Et harum ut harum qui.
    - browser: Et sunt.
      id: Ut cupiditate non.
      ip: Animi ullam.
      isMobile: false
      lastUsed: "2004-07-15T12:58:26Z"
      latitude: Rerum ad.
      location: Atque quod numquam doloribus nihil ut in.
      longitude: Blanditiis perferendis sint quo dolorem provident.
      mapUrl: Dolor omnis.
      os: Eum voluptatem recusandae aut quia non ex.
      userId: Et harum ut harum qui.
    - browser: Et sunt.
      id: Ut cupiditate non.
      ip: Animi ullam.
      isMobile: false
      lastUsed: "2004-07-15T12:58:26Z"
      latitude: Rerum ad.
      location: Atque quod numquam doloribus nihil ut in.
      longitude: Blanditiis perferendis sint quo dolorem provident.
      mapUrl: Dolor omnis.
      os: Eum voluptatem recusandae aut quia non ex.
      userId: Et harum ut harum qui.
...
securityDefinitions:
  api_key_header_API-Key:
    type: apiKey
    description: API Key for users API
    name: API-Key
    in: header
  jwt_header_Authorization:
    type: apiKey
    description: JWT Authentication Security
    name: Authorization
    in: header